Mover userUtils to userModel
